Pro/kharkiv_trolleybus/kharkiv_trolleybus/requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
blinker==1.9.0
cassandra-driver==3.29.2
celery==5.5.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
confluent-kafka==2.10.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.3
daphne==4.1.2
databases==0.9.0
decorator==5.2.1
Deprecated==1.2.18
distlib==0.3.9
Django==5.2.1
django-allauth==65.7.0
django-celery-beat==2.8.0
django-filter==25.1
django-listing==0.10.28
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
dnspython==2.7.0
dotenv==0.9.9
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flower==2.0.1
geomet==0.2.1.post1
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.0
kafka-python==2.2.7
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pika==1.3.2
pika-stubs==0.1.3
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.11.1
pyOpenSSL==25.0.0
pytest==8.3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
PyJWT==2.7.0
redis==6.1.0
service-identity==24.2.0
setuptools==75.8.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tablib==3.8.0
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zope.interface==7.2
